+++ client/src/App.js
@@ -1,14 +1,15 @@
 import { Route, Routes } from "react-router-dom";
-import Home from "./pages/Home";
-
-import Dashboard from "./pages/Dashboard";
-import SignInUp from "./pages/SignInUp";
+import Dashboard from "./pages/DashboardPage";
+import LoginPage from "./pages/LoginPage";
+import BookingPage from "./pages/BookingPage";
+import ConfirmationPage from "./pages/ConfirmationPage";
 
 const App = () => (
 	<Routes>
-		<Route path="/" element={<Home />} />
+		<Route path="/" element={<LoginPage />} />
 		<Route path="/dashboard" element={<Dashboard />} />
-		<Route path="/SignInUp" element={<SignInUp />} />
+		<Route path="/booking" element={<BookingPage/>} />
+		<Route path="/confirm" element={<ConfirmationPage/>} />
 	</Routes>
 );
