+++ client/src/pages/ConfirmationPage.js
@@ -1,0 +1,11 @@
+import { useEffect, useState } from "react";
+import Header from "../components/Header";
+import "./../stylings/Home.css";
+
+export default function ConfirmationPage() {
+	return (
+		<div>
+			<Header />
+		</div>
+	);
+}
